FIX maxValoreConDifficolta

+++ L10/E02/diagonale.h
@@ -13,7 +13,7 @@ typedef struct Diagonale {
 } * diagonale;
 
 bool minorEqualValore(elemento a, elemento b, unsigned int *difficoltaDiagonale);
-unsigned int maxValoreConDifficolta(elemento e, unsigned int difficolta);
+float maxValoreConDifficolta(elemento e, unsigned int difficolta);
 diagonale generaDiagonale(array elementi, unsigned int DD, unsigned int DP, checks controlli);
 uint8_t generaDiagonaleR(array elementi, unsigned int difficoltaDiagonale, checks controlli, link soluzione, uint8_t elementiInseribili);
 bool updateChecks(diagonale d, checks c);
